# specs/001-todo-web-app/contracts/auth.openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend - Authentication API
  description: |
    Authentication endpoints for user signup and signin using Better Auth with JWT tokens.

    **Security Model**: JWT Bearer tokens issued on successful signin.
    All protected endpoints require `Authorization: Bearer <token>` header.
  version: 1.0.0
  contact:
    name: Todo Backend API

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todo-app.example.com
    description: Production server (placeholder)

tags:
  - name: Authentication
    description: User registration and authentication endpoints

paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      description: |
        Creates a new user account with email and password. Passwords are hashed using bcrypt before storage.
        Returns the created user and a JWT token for immediate authentication.

        **Validation Rules**:
        - Email must be valid format (RFC 5322)
        - Email must be unique (returns 400 if already exists)
        - Password must be at least 8 characters
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              valid:
                summary: Valid signup request
                value:
                  email: user@example.com
                  password: securepassword123
      responses:
        '201':
          description: User account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful signup
                  value:
                    user:
                      id: 1
                      email: user@example.com
                      created_at: "2026-01-16T10:30:00Z"
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: bearer
        '400':
          description: Invalid request or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate_email:
                  summary: Email already registered
                  value:
                    detail: Email already registered
                invalid_email:
                  summary: Invalid email format
                  value:
                    detail: Invalid email format
        '422':
          description: Validation error (Pydantic)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Authenticate user and obtain JWT token
      description: |
        Authenticates a user with email and password credentials. Returns a JWT token on success.

        **Token Details**:
        - Algorithm: HS256
        - Expiration: 24 hours
        - Payload includes: user ID (sub), email, expiration (exp)
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
            examples:
              valid:
                summary: Valid signin request
                value:
                  email: user@example.com
                  password: securepassword123
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  summary: Successful signin
                  value:
                    user:
                      id: 1
                      email: user@example.com
                      created_at: "2026-01-16T10:30:00Z"
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: bearer
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Wrong email or password
                  value:
                    detail: Invalid email or password
        '422':
          description: Validation error (Pydantic)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user
      description: |
        Returns the profile of the currently authenticated user based on the JWT token.
        Useful for verifying token validity and retrieving user information.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                success:
                  summary: Current user
                  value:
                    id: 1
                    email: user@example.com
                    created_at: "2026-01-16T10:30:00Z"
        '401':
          description: Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_token:
                  summary: No token provided
                  value:
                    detail: Not authenticated
                invalid_token:
                  summary: Invalid token
                  value:
                    detail: Invalid authentication credentials

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/signin or /api/auth/signup.
        Include in Authorization header as: `Authorization: Bearer <token>`

  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User's email address (must be unique)
          example: user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
          description: User's password (will be hashed before storage)
          example: securepassword123

    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User's password
          example: securepassword123

    UserResponse:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Timestamp when user account was created
          example: "2026-01-16T10:30:00Z"

    AuthResponse:
      type: object
      required:
        - user
        - access_token
        - token_type
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        access_token:
          type: string
          description: JWT token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIiwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiZXhwIjoxNzA1NDk3NjAwfQ.signature
        token_type:
          type: string
          enum: [bearer]
          description: Token type (always "bearer")
          example: bearer

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Invalid email or password

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
          example:
            - loc: ["body", "email"]
              msg: "value is not a valid email address"
              type: "value_error.email"